Restructure SQL compiler
