+++ ibis/backends/base/sql/compiler/__init__.py
@@ -1,0 +1,25 @@
+from .base import DDL, DML
+from .query_builder import (
+    QueryBuilder,
+    Select,
+    SelectBuilder,
+    TableSetFormatter,
+    Union,
+)
+from .transforms import ExistsSubquery, NotExistsSubquery
+from .translator import Dialect, ExprTranslator, QueryContext
+
+__all__ = (
+    'QueryBuilder',
+    'Select',
+    'SelectBuilder',
+    'Union',
+    'TableSetFormatter',
+    'ExprTranslator',
+    'QueryContext',
+    'Dialect',
+    'DML',
+    'DDL',
+    'ExistsSubquery',
+    'NotExistsSubquery',
+)
