+++ ibis/backends/base_sqlalchemy/ddl.py
@@ -3,7 +3,7 @@ import re
 import ibis.expr.datatypes as dt
 import ibis.expr.schema as sch
 from ibis.backends.base.sql import quote_identifier, type_to_sql_string
-from ibis.backends.base_sqlalchemy.compiler import DDL, DML
+from ibis.backends.base.sql.compiler import DDL, DML
 
 fully_qualified_re = re.compile(r"(.*)\.(?:`(.*)`|(.*))")
 _format_aliases = {'TEXT': 'TEXTFILE'}
