+++ ibis/backends/clickhouse/client.py
@@ -11,7 +11,7 @@ import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
-from ibis.backends.base_sqlalchemy.compiler import DDL
+from ibis.backends.base.sql.compiler import DDL
 from ibis.client import Database, DatabaseEntity, Query, SQLClient
 from ibis.config import options
 from ibis.util import log
